requirements.txt
streamlit==1.29.0
openai==1.3.7
requests==2.31.0
pydantic==2.5.2
python-dotenv==1.0.0
tenacity==8.2.3
pytest==7.4.3
requests-mock==1.11.0
anyio==3.7.1
httpx==0.27.2